# src/main/resources/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Health Records API",
    "description": "API for managing patient health records",
    "contact": {
      "name": "Health Records Team",
      "url": "https://healthrecords.com",
      "email": "support@healthrecords.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Default Server URL"
    }
  ],
...
  "paths": {
    "/api/visits/{id}": {
      "get": {
        "tags": [
          "Visit API"
        ],
        "summary": "Get visit by ID",
        "description": "Retrieve a specific visit by its ID",
        "operationId": "getVisitById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Visit"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Visit API"
        ],
        "summary": "Update an existing visit",
        "description": "Modify the details of an existing visit",
        "operationId": "updateVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Visit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Visit"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Visit API"
        ],
        "summary": "Delete a visit",
        "description": "Remove a visit from the records",
        "operationId": "deleteVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": [
          "Patient API"
        ],
        "summary": "Get patient by ID",
        "description": "Retrieves a patient by their ID",
        "operationId": "getPatientById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved patient",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Patient API"
        ],
        "summary": "Update patient",
        "description": "Updates an existing patient record",
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patient successfully updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Patient API"
        ],
        "summary": "Delete patient",
        "description": "Deletes a patient record by ID",
        "operationId": "deletePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Patient successfully deleted"
          },
          "404": {
            "description": "Patient not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/visits": {
      "get": {
        "tags": [
          "Visit API"
        ],
        "summary": "Get all visits",
        "description": "Retrieve a list of all visits",
        "operationId": "getAllVisits",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Visit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Visit API"
        ],
        "summary": "Create a new visit",
        "description": "Add a new visit to the records",
        "operationId": "createVisit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Visit"
                }
              }
            }
          }
        }
      }
    },
    "/api/patients": {
      "get": {
        "tags": [
          "Patient API"
        ],
        "summary": "Get all patients",
        "description": "Retrieves a list of all patients in simplified format",
        "operationId": "getAllPatients",
        "responses": {
          "200": {
            "description": "Successfully retrieved patient list",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patient API"
        ],
        "summary": "Create patient",
        "description": "Creates a new patient record",
        "operationId": "createPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Patient successfully created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visits/search": {
      "get": {
        "tags": [
          "Visit API"
        ],
        "summary": "Search visits",
        "description": "Find visits based on various criteria",
        "operationId": "searchVisits",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Visit"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/patients/search": {
      "get": {
        "tags": [
          "Patient API"
        ],
        "summary": "Search patients by name",
        "description": "Finds patients whose name contains the given string",
        "operationId": "findPatientsByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name string to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved patient list",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Doctor": {
        "required": [
          "name",
          "specialty"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          },
          "specialty": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          }
        }
      },
      "Patient": {
        "required": [
          "dob",
          "gender",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female",
              "Other"
            ]
          }
        },
        "description": "Patient object to be created"
      },
...
      "Visit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "doctor": {
            "$ref": "#/components/schemas/Doctor"
          }
        }
      },
...
      "VisitCreateRequest": {
        "required": [
          "patientId",
          "doctorId",
          "date"
        ],
        "type": "object",
        "properties": {
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}